Working game
I finally managed to make the win conditions work!!
Although, there is a small problem left. If you enter anything else than an Int when choosing the position it will result in an error. I have not yet learned how to catch errors and will probably not implement it in this "game". It was, after all, only a small exercise (that took me hours hahahah)

If anyone reading has suggestions on how to improve my code, then please, feel free to text me :3